+++ node/config.go
@@ -202,6 +202,9 @@ type Config struct {
 	// AllowUnprotectedTxs allows non EIP-155 protected transactions to be send over RPC.
 	AllowUnprotectedTxs bool `toml:",omitempty"`
 
+	// EnableDoubleSignMonitor is a flag that whether to enable the double signature checker
+	EnableDoubleSignMonitor bool `toml:",omitempty"`
+
 	// BLSPasswordFile is the file that contains BLS wallet password.
 	BLSPasswordFile string `toml:",omitempty"`